Initial commit: Add folder structure + 3 PoCs

+++ pocs/sandwich-attack/src/SandWichAttack.sol
@@ -5,29 +5,26 @@ import "forge-std/Test.sol";
 import "forge-std/console.sol";
 
 interface IUniswapV2Pair {
-    function swap(uint amount0Out, uint amount1Out, address to, bytes calldata data) external;
+    function swap(uint256 amount0Out, uint256 amount1Out, address to, bytes calldata data) external;
     function getReserves() external view returns (uint112, uint112, uint32);
 }
 
 interface IERC20 {
-    function balanceOf(address) external view returns (uint);
-    function transfer(address, uint) external returns (bool);
-    function approve(address, uint) external returns (bool);
-    function transferFrom(address, address, uint) external returns (bool);
+    function balanceOf(address) external view returns (uint256);
+    function transfer(address, uint256) external returns (bool);
+    function approve(address, uint256) external returns (bool);
+    function transferFrom(address, address, uint256) external returns (bool);
 }
 
 interface IUniswapV2Router {
     function swapExactTokensForTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
-    ) external returns (uint[] memory amounts);
-    
-    function getAmountsOut(
-        uint amountIn, 
-        address[] calldata path
-    ) external view returns (uint[] memory amounts);
+        uint256 deadline
+    ) external returns (uint256[] memory amounts);
+
+    function getAmountsOut(uint256 amountIn, address[] calldata path) external view returns (uint256[] memory amounts);
 }
